:recycle: Move stock chart options to ChartConfigService
